requirements.txt
Flask
gunicorn
flask_talisman
psycopg2-binary
python-dotenv
Flask-SQLAlchemy
config
pytz